Allow for specifying the organization id in the configuration (#364)
Allow the organizationId to be specified in the configuration. See
https://platform.openai.com/docs/api-reference/organization-optional

Right now there is no way to specify this.

I added a `.organizationId(System.getenv("OPENAI_ORGANIZATION_ID"))` to
the builders in a bunch of the tests. They should have no effect other
than if you want to run the tests to pass an organizationId, you can
just set the `OPENAI_ORGANIZATION_ID` environment variable. The
`OpenAiClient` won't do anything if the `organizationId` is `null`.

Happy for questions/comments.

Fixes #344