Clarify the mandatory and optional Azure connection parameter (#636)
Based on https://github.com/langchain4j/langchain4j/pull/635 we can see
that the `serviceVersion` is not mandatory. If not set, then the latest
version of the API is used. So only `apikey` and `endpoint` are really
mandatory. Then, is `serviceVersion` and `deploymentName` are not set,
we set default values.