Implement remove methods for InMemoryEmbeddingStore (#1220)
## Issue
[https://github.com/langchain4j/langchain4j/issues/301](https://github.com/langchain4j/langchain4j/issues/301)

## Change
I've implemented remove methods for InMemoryEmbeddingStore.

Current problems:
1. I'm not completely sure how to test all of this. And, as I've seen,
you don't fully test it too)
2. There is a very nasty problem with `removeAll(Filter filter)` . Let
me try to break it in small pieces:
1. `EmbeddingStore`'s are parametrized with `Embedded` type parameter.
2. `removeAll(Filter filter)` method in `EmbeddingStore` accepts
`Filter`.
   3. `Filter` accepts `Metadata` object.
   4. `Metadata` object is stored in `TextSegment`.
5. `TextSegment` is usually what is `Embedded`. But it's not guaranteed.

How my pull request deals with this problem: 
1) for every entry in `InMemoryEmbeddingStore` check the type of
`Embedded`.
2) If it's a `TextSegment`, then proceed.
3) If it's not, then raise `UnsupportedOperationException`.
The issue with this strategy is that we perform the check for every
entry. I haven't found a way in Java to check the type of type parameter
(because it's erasured in runtime). (*In C++, for example, there are
`constexpr if`, `<type_traits>`, `std::is_same<U, V>`, explicit
specialization, etc.*)

We could solve this problem if `InMemoryEmbeddingStore` would support
only `TextSegment` as `Embedded`, but that's a breaking change.

I could also not implement this method, but, I really really need it in
my project (if we chose to stick with `InMemoryEmbeddingStore`).
  

## General checklist
<!-- Please double-check the following points and mark them like this:
[X] -->
- [x] There are no breaking changes
- [ ] I have added unit and integration tests for my change
- [ ] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [ ] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green
<!-- Before adding documentation and example(s) (below), please wait
until the PR is reviewed and approved. -->
- [ ] I have added/updated the
[documentation](https://github.com/langchain4j/langchain4j/tree/main/docs/docs)
- [ ] I have added an example in the [examples
repo](https://github.com/langchain4j/langchain4j-examples) (only for
"big" features)