Fix DefaultPromptTemplateFactory RegExp for Android. (#1518)
## Issue
Closes #1075 #1212

## Change
Adjusted the regex according to the discussion in both issues. Tested
the change with a sample repository using an instrumentation test which
uses the DefaultPromptTemplateFactory. The test can be found
[here](https://github.com/marble94/langchain4j-android-regex-fix/blob/5640a9743e33f7359305596fccf0fe4cb816a54a/app/src/androidTest/java/com/example/myapplication/ExampleInstrumentedTest.kt#L20).

If a test is required in this repo please advice me on possible steps,
I'm not sure how to test it without setting up an Android environment.

## General checklist
<!-- Please double-check the following points and mark them like this:
[X] -->
- [x] There are no breaking changes
- [ ] I have added unit and integration tests for my change
- [x] I have created an example Android repository verifying that the
change fixes the runtime error in Android
- [x] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [x] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green
<!-- Before adding documentation and example(s) (below), please wait
until the PR is reviewed and approved. -->
- [ ] I have added/updated the
[documentation](https://github.com/langchain4j/langchain4j/tree/main/docs/docs)
- [ ] I have added an example in the [examples
repo](https://github.com/langchain4j/langchain4j-examples) (only for
"big" features)
- [ ] I have added/updated [Spring Boot
starter(s)](https://github.com/langchain4j/langchain4j-spring) (if
applicable)