660 : Support for onRetrieved for TokenStream (#1527)
## Issue
Closes #660 

## Change
Refactor of `TokenStream` interface and `AiServiceTokenStream` class so
that all methods available in sub interfaces are represented in the
`TokenStream` interface instead. Makes it possible to call methods in
any order. This has the side effect that a developer can now call e.g.
`onNext` two times on the same instance of a `TokenStream` and the last
call will be used.

The rationale for this refactor was that the interface
`OnCompleteOrOnError` did not fit well when trying to add the new
`onRetrieved` callback handler. This will make it possible to more
easily add more features in the future.

Added new callback method `onRetrieved` which is called **before**
invoking the language model in `AiServiceTokenStream`, no changes are
necessary in current model integrations.

I have not yet deleted the `OnCompleteOrOnError` and related interfaces
but they have no use in the code anymore.

This is a first attempt and any feedback is appreciated.

## General checklist
<!-- Please double-check the following points and mark them like this:
[X] -->
- [X] There are no breaking changes
- [X] I have added unit and integration tests for my change
- [X] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [X] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green
<!-- Before adding documentation and example(s) (below), please wait
until the PR is reviewed and approved. -->
- [ ] I have added/updated the
[documentation](https://github.com/langchain4j/langchain4j/tree/main/docs/docs)
- [ ] I have added an example in the [examples
repo](https://github.com/langchain4j/langchain4j-examples) (only for
"big" features)
- [ ] I have added/updated [Spring Boot
starter(s)](https://github.com/langchain4j/langchain4j-spring) (if
applicable)